Refactor ConnectCheck method
Merge the horizontal and vertical methods, combine them into one method and add some common variables as values that need to be replaced when checking different directions.
Add a new enumeration of direction as the judgment parameter for using ConnectCheck(), and realize multi-loading, so that user can choose which direction to judge freely, or directly call the original ConnectCheck() to judge the whole direction.

Translated with www.DeepL.com/Translator (free version)