Refactor ConnectCheck() method
Added "nodeIdXBegin, nodeIdYBegin, nodeIdXEnd, nodeIdYEnd, nextX, nextY"
and simplify the method.
These variables allow the checking and checked pieces to input XY coordinates freely, instead of using fixed coordinates, which will be more helpful for Slash and Backslash determination.
Therefore, the extra switches for determining the coordinates of the current checking and checked pieces are also removed, and the decision is made directly at the beginning.