Add Slash and Backslash Connection
Add Slash and Backslash min and max case to ConnectRangeConverter().
Add Slash and Backslash case to ConnectCheck(), and remove old case.
Fix bug of X and Y of nodeIdEnd  of Horizontal case of ConnectCheck().
Refactor for loop of ConnectCheck(), and add new condition for horizontal and vertical condition.
Add new variable name jCount, that use to count j for loop times.
Add new variable name nextXZeroPlus and nextYZeroPlus, that use to avoiding non-entry for loops.